backend/requirement.txt
certifi==2018.8.24
Django==1.8
django-rest-framework==0.1.0
djangorestframework==3.8.2
Markdown==2.6.11
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
